fix(KIJK-203): fix redirect error after login/ signup
- add error message on auth page
- better lazy loaded auth pages
