+++ apps/client/src/app/auth/login.tsx
@@ -1,0 +1,49 @@
+import { Dispatch } from 'react';
+import { RouteApi, useNavigate } from '@tanstack/react-router';
+
+import { UserAuthForm } from '@/app/auth/auth-form';
+import { Icons } from '@/components/icons';
+import { Button } from '@/components/ui/button';
+import { useToast } from '@/hooks/use-toast';
+import { useAuthStoreActions } from '@/stores/auth-store';
+
+const route = new RouteApi({ id: '/auth' });
+
+export function Login({ goto }: { goto: Dispatch<React.SetStateAction<'Login' | 'Sign Up'>> }) {
+  const { login } = useAuthStoreActions();
+  const { toast } = useToast();
+  const navigate = useNavigate();
+  const search = route.useSearch();
+  const from = search.from ?? '/';
+
+  const handleLogin = async (email: string, password: string) => {
+    const result = await login(email.toLowerCase(), password);
+
+    if (result.error != null) {
+      toast({
+        title: 'Something went wrong.',
+        description: result.error.message ?? 'Your sign in request failed. Please try again.',
+        variant: 'destructive',
+      });
+    }
+    void navigate({ to: from, params: true });
+  };
+
+  return (
+    <div className='container flex h-screen w-screen flex-col items-center justify-center'>
+      <div className='mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[400px]'>
+        <div className='flex flex-col space-y-2 text-center'>
+          <Icons.logo className='mx-auto h-12 w-12' />
+          <h1 className='text-2xl font-semibold tracking-tight'>Welcome back</h1>
+          <p className='text-sm text-muted-foreground'>Enter your credentials to log in to your account</p>
+        </div>
+        <UserAuthForm onSubmit={handleLogin} btnLabel='Login' />
+        <p className='px-8 text-center text-sm text-muted-foreground'>
+          <Button variant='link' onClick={() => goto('Sign Up')}>
+            Don&apos;t have an account? Sign Up
+          </Button>
+        </p>
+      </div>
+    </div>
+  );
+}
