+++ apps/client/src/routes/_protected/home/index.component.tsx
@@ -9,10 +9,6 @@ import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 
-// export const Route = new FileRoute('/_protected/home/').createRoute({
-//   component: HomeIndexPage,
-// });
-
 export const component = function HomeIndexPage() {
   return (
     <>
